refactor: shortHands=>universalShorthands
