+++ packages/pnpm/src/shorthands.ts
